+++ activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb
@@ -249,6 +249,10 @@ module ActiveRecord
         true
       end
 
+      def supports_uuid?
+        true
+      end
+
       def supports_comments?
         true
       end